requirements.txt
requests>=2.25.1
python-dotenv>=0.19.0
pydicom>=2.3.0
python-dateutil>=2.8.2